+++ project_management/agile/serializers.py
@@ -10,3 +10,9 @@ class ValueSerializer(serializers.ModelSerializer):
     class Meta:
         model = Value
         fields = '__all__'
+
+
+class PrincipleSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Principle
+        fields = '__all__'