+++ project_management/agile/views.py
@@ -18,3 +18,15 @@ class ValueDetailAPIView(generics.RetrieveUpdateAPIView):
     queryset = Value.objects.all()
     serializer_class = ValueSerializer
     permission_classes = [AllowAny]
+
+
+class PrincipleListAPIView(generics.ListCreateAPIView):
+    queryset = Principle.objects.all()
+    serializer_class = PrincipleSerializer
+    permission_classes = [AllowAny]
+
+
+class PrincipleDetailAPIView(generics.RetrieveUpdateAPIView):
+    queryset = Principle.objects.all()
+    serializer_class = PrincipleSerializer
+    permission_classes = [AllowAny]